fix: disable CSP for now
Currently the frontend is still making "generic" requests. EG, for
prometheus. Longer term, I'd like to shift this over to Rust. This would
be more performant, and we can just push updates to the frontend.

It would also mean we can have a super strict CSP, and only make generic
requests for certain things from Rust.

We're not there right now though.