feat: sort runbooks sort-runbooks-alphabetically (#212)
I think this might make sense as a new default. Sorting by updated might
be weird if you have a runbook you refer to often, but rarely edit. At
least with alphabetical, the order is easily predicted.

In the longer term, we'll likely want some sort of toggle for changing
this