fix: Remove mutable Y.Doc from Runbook model (#224)
* Remove mutable Y.Doc from Runbook model

The `ydoc` field on the Runbook model has always stored a YJS update created from
`encodeStateAsUpdate()`. Previously, we used this update to hydrate a fresh copy
of the complete `Y.Doc` upon loading the model from the database. However, since we're
treating models as immutable, the editor could accidentally operate on an old copy of
the `Y.Doc` instance.

This commit changes this behavior, keeping the data in the modal as *only* the static
`Uint8Array` representing the state-as-update, and we only hydrate into a mutable
`Y.Doc` instance in the `Editor` component, where it is used in the editor and in the
synchronization provider.\