refactor: move pty_sessions into its own store (#423)
* refactor: move pty_sessions into its own store

This also move the ownership of all data into a tokio task, and uses
message passing to manage the state. This means we don't need to have
locking ALL over the place, and can instead communicate a little more
cleanly.

Given that there is message passing from the frontend -> backend
already, this seems to make the most sense.

* fmt

* clippy