Improve Phoenix channel joining (#458)
Phoenix channels throw an error if you attempt to `join()` them more than once.
Unfortunately, the docs aren't super clear on how to determine if you've already
joined. The channel object provides a `state` property that should, in theory,
allow us to check this; we use this in `ensureJoined`, but we don't do any
checking in `join` itself.

Thus, this change makes `join()` private, making `ensureJoined()` the public
entry into the process of joining a channel. If we're already joined, or
in the process of joining, we return a cached promise from the original
`join()` call. Since channel joins can return data, `WrappedChannel` is
now generic over a type called `J` that represents the type of this data.