fix: ssh block border updates (#466)
* fix: ssh block border updates

We needed events :(

I tried hooking into the global editor.onChange but had weird issues.
Hence for now, we are emitting the block create/delete events from the
useEffect hook of the blocks we are currently interested in.

Should probably

1. make editor.onChange work?
2. add the event emitter to all blocks :/

As discussed on slack, I would like to sort some sort of global event
hub. This isn't it - I don't want to make new buses for every type, it
feels wrong.

* toast errors, default port

* wip

* build

* correct hook usage in the terminal too