fix: blank runbook editor (#526)
I had an error reported where the runbook page was blank. At around the
same time, we had DESKTOP-FRONTEND-6R trigger on sentry

```
Could not find snapshot based on tag: latest
```

I suspect there is a deeper issue here, with a race condition or
similar. However, imo regardless it makes sense to fallback to latest in
the case of a missing tag, rather than fail to load at all.