fix: mysql rendering latin1_bin table varchar as binary (#549)
Upstream isseue https://github.com/launchbadge/sqlx/issues/3387

SQLx mysql driver incorrectly interprets varchar inside a latin1_bin
table to be varbinary. We have no way of interpreting this from the
client.

Users still expect to be able to read the text. So, if most characters
in the string are printable, then it's probably binary